more dialogue troubleshooting
issues:
as a test, i attempted to have 2 forms of interaction. one is "talking", which is done via the now context-sensitive attack button. NPCs cannot be attacked, so instead, the attack button talks with them.
the second is "examine", which is done via its own button. everything is examinable, including enemies.