ALMOST working properly
the player now takes damage from a single hitbox, as the damage function is now built into the 'hit' animation.

i noticed an untinded bug/feature: when the player is hit, invicibility begins, lasting for 1 second. when the player rolls, a shorter invincibility time happens. the issue is that the roll invinc.time overrides the hit invinc.time, which could cause gameplay issues. basically, i don't want the player to be able to be hit more than once / second.