working on charge attack
i feel like i am going about this the totally wrong way, but if i can get it to kind of work, then i will try to reprogram it properly.

the first thing i did was refactor the combo attack code for easy consecutive attacks.
what i want to accomplish is if the player presses the ROLL button after having held the attack button for 1 second, they perform a much stronger attack. the player will turn red indicating the necessary amount of time has elapsed. if the player releases the attack button without pressing roll, their state and colour goes back to default.

some issues right now are sometimes the player gets stuck in the "charged" state. sometimes they enter the charged state without the button being held. this is probably because i am using a timer instead of using delta.

the next issue is that the strength modifier (+4) doesn't stay long enough on the attack animation to take effect, i think. i was going to change this when i made a new attack animation and attaching the method there, which i don't think will be a problem, but i'll try and get it to work in code first.