working so far, just need to figure out a couple issues
one i knew i was going to have to deal with was getting hit during the special attack animation. the beginning of the animation increases the sword's strength, but if the player is hit during the animation, it never reaches the end to revert the sword's damage back to the orig value. will either have to build something into the hit function to make sure that happens, or in the shade state. i'm thinking of making the shade attack totally invulnerable, as well. if i do that, this ceases to be a problem.