issues
the problem when anchoring the chargeUI element to the player is that it was drawing behind the bushes and other things in the YSort.
so i tried moving it to the GUI canvas layer, but i couldn't get getting player position to work properly.
so i tried moving it as a child of the root world node, and now it looks like it's stuttering when it's moving.