it is literally impossible to move the charge gauge to where i want it to appear on the screen
this is the most nonsensical, fucking waste of time i've ever dealt with. nothing is predictable, nothing works, the charge gauges always draws below the bushes when i get close to where i want it. wherever i move the anchors for either the control node or the texture, it moves in COMPLETELY UTTERLY random ways. i've been trying to figure this out, but i've been wasting my time, because it's a piece of fucking garbage.