adding sound effects to item pickup, issue
and once again, a nonsensical issue arises: the coin and heart collect sound effects, when played, have a audible glitch that occurs when the audio playback ends. this does not occur with other sound effects, even when called in the same way.
there is no audio glitch in any of the sound files. i have tried extending the length of the sound files with silence, but the glitch still happens when the duration is reached, but only with the newly created coin and heart sound effects. for example if i make the "enemy death" sound play instead of the coin collect sound, it plays fine with no glitch.