item collect sound effect glitch fixed
potential issue: in order to prevent too many collect sounds from being played at once, each audio file has their own AudioStreamPlayer in the root node, World. The outcome is that while both heart and coin sound effects can be played simultaneously, multiple instances of either heart or coin sound effects will not be played simultaneously.
i figure i will have to do this for items that there could potentially be many of on screen at one time.

another way of going about this, which i tried, was to instance AudioStreamPlayers in the ItemCollectEffect scene itself. the issue is that multiple ASPs for the same sound effect could be instantiated at once if the player collects a lot of coins, for instance, in a short amount of time. the sound effect, played multiple times all at once, would produce unwanted clipping in the audio (distortion). right now, instead of instantiating new ASPs for each sound effect, it tells the 2 sound effect players in the world scene when to play.