uploaded crow asset/created crow class
i have copypasted the code from the bat class, and will be rewriting it.
things that will differentiate from the bat:
the crow has 3 distinct animations: takeoff, flight, and landing, in contrast to the bat's 1. when the crow is in the idle state, it should play the idle animation; when it wants to move from the idle state, it should takeoff and then play (loop) the flight animation until it wants to land, in which case it should play the land animation and re-enter the idle state.
another thing i want to attempt is to give the crow a distinct 'attack' where its hitbox is active. ie, simply touching the crow should not damage you; it must perform its attack animation and hit you with it.