created attack zone for bat / cloud shadow works but is ugly
i will first attempt an attack state for the bat, as the bat is more or less complete and functioning properly otherwise.
i'm not sure how i want to go about doing this...
currently, there is a player detection zone. upon entering this zone, the bat will begin moving towards the position of the player (CHASE state). the bat hitbox is always enabled.
maybe, once the bat gets within a certain range from the player (AttackPlayerZone), it enters the ATTACK state during which it stops moving briefly before unleashing a thrust attack, during which the hitbox will be enabled. the bat will then move towards the player in an attempt to hit them, then revert back to the position they were at when the player entered the bat's AttackPlayerZone. the state is reset to CHASE and behaviour continues until one of 3 things happen:
the player enters the AttackPlayerZone again (state CHASE -> ATTACK);
the player leaves the PlayerDetectionZone, (state CHASE -> IDLE or WANDER) or;
the bat dies ( -> state DEAD).
theoretically, if i do it this way, the only new code added will be for the ATTACK state. i will either need to set up timers to determine for how long the ATTACK state lasts, or call methods from an animation player, or both.