working on enemy attack behaviour; issue.
# attacking
# the attack_player() function is run when the enemy is in the CHASE state
# if that player enters the enemy's attackPlayerZone AND the attack is NOT on cooldown:
# 1. detection is disabled;
# 2. attacking is set to true;
# 3. enables the enemy's hitbox and quadruples the sprite animation speed
# 4. changes the state to ATTACK and starts the attackTimer (0.8s)
# 5. ATTACK state: gets the player's position point before setting attacking to false
# 6. accelerates the enemy to the player's position point
# 7A. if the enemy arrives at the point, resets the sprite animation speed and state to IDLE
# 7B. on attackTimer_timeout, attack_on_cooldown becomes true
#     disables hitbox
#     resets sprite animation speed and state to IDLE
#     starts a 1s timer, after which attack_on_cooldown becomes false
#     enemy detection is re-enabled

tldr: attack state lasts for 0.8, after which the attack is on cooldown for 1s.
ISSUE: i want that attack cooldown to be interruptable by the player's attack, however hitting the enemy while attack_on_cooldown is true seems to break the AI. it re-enables the detection zones, but the player ceases to be recognized by them ONLY if the attack cooldown is interrupted, turning interrupted enemies into harmless vegetables.

also, added a Test world where the player can 1v1 enemies to test their AI.