enemy AI working as intended... i think
fixed the previous issue of interrupting the enemy's attack cooldown breaking its AI. it seems to work properly now.
an added benefit gameplay wise is that the bat's velocity during the attack state is increased so much that the sword's knockback velocity doesn't effect it enough to knock an attacking bat safely away. therefore, attacking bats must be dodged instead of simply spamming attacks that they mindlessly run into.
also, since the attackPlayerZone is so small, the player can't start an attack fast enough to hit a bat that's already started attacking. it takes 0.1s for the sword hitbox to activate after beginning the attack, and the bat contacts the player in less time than that.