working on making Home objects interactable
the interact/talk logic is a bit strange. talk logic exists on NPCs that you can talk to and Interact logic exists on the Player. the reasoning behind this is that while all interactable objects are interactable, not all interactable objects are talkable. so the idea is to have talkable objects handle their own dialog, and have the player handle all interactable ("examinable") objects themself. i have a feeling this will bite me in the ass later but it works so far.