interaction tweaks
context:
the contextual interaction check does not check if talkTimer is_stopped, therefore multiple interactions can be made in a row. in other words, the interaction box is not disabled. however, the talkTimer is started by interaction; therefore, on its timeout, it re-sets noticeDisplay(s) as necessary.
issue:
the examine notice does not disappear, which is fine, however on talkTimer timeout - if there is an unexamined object in range, it plays the notice sound effect, which i don't want it to.
possible solutions:
1. have the interaction behave identically to talking, in that you can only interact once every 0.5s. probably the simplest solution, but not fully desirable. cons: can't interact multiple times in a row.
2. have the set_notice function check if what started the talkTimer was an Interaction - and if it was, forego playing the sound effect. con: possibly needlessly complex.
3. instead of interacting starting the talkTimer, it has its own timer which does not run the set_notice functions on timeout. con: will need more timer checks where necessary.