interaction working as intended
as a challenge to myself, i gave the desk 2 additional states of "examined", whether its lit or unlit. if you examine it while unlit, the notice will disappear - but it will reappear when you turn the light on, indicating alternate dialog.
if you turn on the lamp before examining it, you can exhaust its examine dialog, then the notice will disappear. however, if you switch the lamp back off, the notice will appear again - ONLY IF you didn't examine it before turning it on.
the basic intention is: if there is unseen dialog from examining, the notice will show up - however it won't show up if an object needs to be interacted with before its new dialog is encounterable.
basically: examine everything, then interact with it, then examine it again, if you wan't to get 100%.