lightswitches added for both rooms
context:
each interactable object has an examined property. if an object is "examined", the "?!" will not pop up for it anymore.
issue:
there will be many duplicates of examinable objects, which means each new instance will always trigger the "?!" notice display.
solution:
create a singleton script that will keep track of every examinable object in the game each with their own "examined" property, which will then be referenced regarding whether the noticeDisplay pops for that particular examinable object or not.