basic tumbleweed working
description:
the TumbleweedSpawner instantiates new Tumbleweed scenes 1 screen (320px) to the right of the player.
on the X-axis, the Tumbleweeds have a base MAX_SPEED of 96, and a random modulator of 32 (effective speed of each individual tumbleweed varies between 96 and 128). currently, this speed is set on spawn and fixed for the lifetime of the tumbleweed. if changing weather effects like wind is to be added, this number could potentially be modified further.
Y-velocity changes depending on whether the player is moving or not. if the player is not moving on the Y-axis, the Y-velocity is set between -16 and 16. if the player IS moving, the Y-velocity is set between 1/8th of the player's Y-velocity and the player's Y-velocity, clamped to max out at 80 (or -80).
issues: animation & sound
the animation has no "bounce" to it, which will need to be added.
there is only 1 tumbleweed sound. more will need to be added from which 1 will be randomly chosen when a sound is to be played (on "bounce").
also, the movement itself does not feel right. perhaps correcting the animation will correct this issue. animation should be fixed before further modification to movement behaviour is made.
overall still needs a lot of tweaking in terms of speed, direction, spawn time, and spawn location.
the next big step is to add changing winds that effect both the sandstorm overlay as well as the tumbleweed movement. this wind may also hinder player movement.
i want to make it a challenge simply to traverse the desert. things like too much sand blowing in your face or spending too much time in the sun should cause negative status effects, which can be cured by visiting an oasis, for example.