refactored some player stats into PlayerStats scene as well as status setting
there are a number of ways to go aboud handling status effects. probably the best way but one i don't know how to do would be to use a dictionary.
for now, i'm using a setget in the PlayerStats to handle when status gets applied, and using modulator variables to modify the original values.
i also had the idea of having each status effect as its own scene. this might be a bit more complex but may end up being cleaner in the end. i have to continue working with status effects to know how i want to handle this in the future.