run animation is slowed during slow status effect
i had to replace the Run BlendSpace2D node in the AnimationTree with a BlendTree, then re-create the Run BlendSpace2D inside of the BlendTree. the animation then goes through a TimeScale node, the scale of which is changed through code.
when the PlayerStats recieves a new status, it runs the set_status(value) function and emits the status_changed signal. the value here is the key for the status, which is then matched with the appropriate behaviour...
in this case, "slow" makes the speed_mod = -50, and "not_slow" makes the speed_mod = 0. the speed_mod and max_speed are added to determine the player's actual movement speed.
as mentioned, a signal is also emitted, which is connected to the Player scene (line 94), which runs the apply_status Player function. currently, this function determines the TimeScale of the Run node in the AnimationTree ("slow" = 0.5, "not_slow" = 1).
simply put, the PlayerStats scene handles the mechanics of status changes, and the signal emitted allows the Player scene to modulate various animation parameters locally. this will potentially be useful for animating other status effects.
if there is to be a Poison status effect, for example, the PlayerStats scene would handle HP loss presumably, and the signal recieved by the Player scene would play a specific animation for being poisoned.