added poison status effect
press 'T' to test the poison effect.
made a Poison scene for the PlayerStats. Poison is a status which deals damage after an elapsed period (a "tick"). each period is currently 3 seconds, set by the Poison scene's Timer.
the Poison has 3 variables: count, duration, and potency.
count: poison tick index.
duration: number of total ticks.
potency: the damage of each tick.
example of functionality: normally - when the player is not poisoned - upon contracting the status, the count is incremented to 1; then, the timer times out, damage is incurred, and the count is incremented to 2; timeout, damage, count = 3; then finally timeout, damage, and removal of the status since count = duration, 3.
there is an exception, which is if the player contracts poison while already poisoned: in this case, the count is set to 0 instead of 1. the logic here is that while Poison can not stack, its count is reset when the status is reapplied only AFTER the current tick, meaning a fresh poison is "queued" to start following the current tick.
in practice:
contract poison once: 3 total damage over 9 seconds.
contract poison, then contract poison again before the first tick of damage: 4 total damage over 12 seconds.
contract poison, 1 tick passes, then contract poison again: 5 total damage over 15 seconds, etc.