fixed tallGrass error/naming convention; starting work on PlayerLog
PlayerLog is an autoloaded singleton node meant to keep track globally of things the player has done. right now, its purpose is to keep track of whether objects in the world have been fully examined or not. if an object becomes fully examined, it's state is changed in the PlayerLog. upon entering an area and spawning all objects, on each examinable object a check is made to the PlayerLog to see if it has been fully examined (globally). if so, sets the examined property on the object locally.

theoretically, this system should work with repeatable objects, ie. enemies. if an enemy type is fully examined, all enemies of that type should become fully examined. typing that sentence out has revealed an issue: the check to the PlayerLog is only made when the map is loaded, meaning it won't automatically update other objects of the same type locally after one object becomes examined. thus, a signal will have to be sent from the PlayerLog when a property is modified, to be recieved by every examinable object, changing its local examined property from false to true. i think.