using signals in the PlayerLog singleton seems to work OK
i'm not sure how expandable this code is, however.
plan:
objects of which there are multiple will have setget functions in the PlayerLog scene. when an examined property is set in the PlayerLog, a signal is emitted for that object type. all objects of that type will connect to this signal on their _ready() functions. this signal will run an "examine_complete" function, which sets the local examined property of the object to true.
some of this code is probably redundant. i can probably re-do it simpler without the setgets using signals only. will look into that solution