previous solution worked
got rid of the setgets. the code is much simpler now:
when examining an object for the final time, runs a function in the PlayerLog corresponding to the examined object that both sets the global examined variable to true, and emits a signal that all objects of the same type listen for. this signal runs a function that sets the local examined variable to true.