applying previous logic to objects' dialog index
translation: the objects should share a dialog index. for example, if one of 2 instances of the same object that has 2 lines of dialog is examined,  the first line of dialog is displayed. if then the other instance of the object is examined, the second line of dialog is displayed.
the PlayerLog is now responsible for advancing the dialog index in the function set_dialog_index(name, value) where 'name' is the name of the object and 'value' is the index to which it advances. this function emits a signal which the object listens for. when that occurs, the object's advance_dialog_index(value) function is called, with 'value' corresponding to the dialog index to which the object should advance.