programming challenge with the lightswitches
i wanted to experiment with how different scenes and their variables can interact with each other. the lightswitch and how it is examinable is pretty complicated.
first, if neither switch has been switched on, examining either one will produce the message "A switch of some kind". both will give the examine notice. on first examine, the examine notice will disappear for both lights. this is the only dialog before turning either of the switches on.
when a light is turned on, the notice comes back for both switches, indicating new dialog. the first line for both on & off positions is "A lightswitch." if a lightswitch is examined a second time, a message will display whether it is in the 'ON' or 'OFF' position.
for the examine notice to permanently disappear, one must see both the 'ON' and 'OFF' dialog boxes. otherwise, toggling either switch, regardless if its new state has been examined or not, will re-instantiate the examine notice, for both switches. seeing both 'ON' and 'OFF' dialog for either switch will turn off the examine notice for both.