inventory almost working
GameManger attemps to load user://inventory.tres;
if not found, the game creates inventory.tres file and gives the player 3 potions, which is saved/loaded to/from the inventory.tres file
going to work on a UI inventory display to test setting items in real-time.
plan:
will probably have a DarkSouls-style inventory, where you can scroll back & forth between your current items, and then assign a "Use_Item" input to a key, which will run a function that will use the current item in the slot.
if it is a stackable consumable, Use_Item will activate its effects (animation?) then will decrement its quantity by one. if there is only one left, it will delete it from the inventory.
if it is not a consumable, it will display a message saying you cannot use that here.