bugfixes & issues
inventory works at keeping track of items in real-time, and also can save to file (though i've removed this feature until i figure out how i'm going to handle saving progress of all other game data).
stackable items stack properly and a new stack is generated when an item's max stack is reached. however i'm thinking of removing this feature and instead keep a maximum of 1 stack of a particular item type, rather than give the player the option to fit every inventory slot with a stack of potions.
to do:
the next important thing to figure out is removing items from the inventory.
also on picking up an item, a notice should appear, indicating it's been placed in the inventory. the inventory UI should display a sprite of the current slot's item.
since originally the player's inventory was saved to a file, this means i will have to keep track of the inventory another way.
currently, the Exit script takes the exiting player's inventory and stores it in variable new_inventory, which is passed as part of the attributes argument in the global goto_scene function. the player is then re-instantiated.
after hours of debugging, i cannot seem to reconnect the GameManager's initialize_player() function to the inventory UI. it does so successfully on first instantiation, ie. when the program starts. but when the player gets reinstantiated in a new scene, no methods i've tried seem to work to reconnect the UI. it disappears. i am going fucking insane trying to fix this. i have made 0 progress for 3 fucking hours, i have 0 fucking errors, and there is zero fucking discernible reason as to why this doesn't fucking work. the inventoryUI's _on_player_inventory_changed function simply stops running. it runs once. then it doesn't ever again. despite creating a separate "reinitialization" signal. despite making another connection there. what fucking trash. fucking garbage waste of fucking time.
connecting the GameManager's reinitialize_player function to a key, works however! fuck this fuck this fuck this. it's not fucking complicated to get a function to run. you tell it to fucking run. i've told it to fucking run in so many fucking ways, and it won't. it won't even err. it's infuriating. but when i bind it to a fucking keystroke it works. i don't it to be hidden by a keystroke, i want it to run like a fucking computer program should. why the fuck doesn't the player get fucking reinitialized when i fucking tell it to? so many fucking times?