working on inventory; will start from scratch
the inventory code has becomes almost too convoluted. i went for a different programming approach and tried to implement something that was to be scaleable and customizable, but its too unwieldy. it's a total mess. the item class, GameManager and UI are relatively fine. but the actual inventory script is all over the place. an inventory system like this would work for a big RPG with lots of different kinds of items, not for a simple action RPG where there will only be a small handful of usable items.