sink interactable / audio edits / bugfixes
sink now has working sound and animation.
need to work out how i'm going to handle items. the metal_pot will be the challenge: using the metal pot on the sink while it's running will change it to a filled_metal_pot. then, using the filled_metal_pot on the stove will place it on the stove (i haven't programmed the stove yet). the player should be able to do this regardless of if the stove is turned on or not. if it is not turned on, the pot is simply placed there and nothing happens. if the stove is turned on, the water in the pot will begin to boil.

i occasionally encounter 2 bugs with player attacks:
one is the sword hitbox not getting switched off if the player is hit during an attack. this is difficult to replicate, and generally happens when the player gets hit while spamming attacks. i've tried to ensure the sword hitbox stops monitoring on hit, but something else may have to happen as well.
a less game-breaking bug is sometimes, the charge sound plays when it shouldn't. more accurately - it's possible that the charge sound is not STOPPED when it SHOULD. this usually occurs when either interacting with other objects (changing what the attack input does), or when out of stamina (when charging shouldn't be possible).
it might be that the sound gets played and stopped on the same frame, but it is called to stop before the sound begins playing, so the charge sound plays out despite no charge taking place.
a more game breaking bug i encounter very rarely is the player freezing. this can only mean that the player gets trapped in a state transition - meaning, a new state is supposed to applied, but isn't. player input is only listened for during the move state, so the ultimate issue is sometimes the move state is not re-set after moving quickly between other states. eg. getting hit from the attack state means the attack state is interruped and changed to the hit state. at the end of the hit animation, a function is run during which the state is changed back to MOVE, so i'm not sure why the state change wouldn't be called.