working on item usage
ItemHandler should handle all item logic, ie. what the item does. inventory only manages checking for items, updating them, and removing them.
when the Player enters an object's InteractBox, checks if the object has the variable "item_usable". if it does, the player's "using_item" variable is set to true.
to do:
ItemHandler needs to know if the object on which an item is being used is the correct object for that particular item. if it is, ask the player if they wish to use the item (to confirm), and if not, display a default message ("cannot use item here").
if "using_item" is true and the used_item.type != "Consumable",