item matching working (i think)
when approaching an object with the item_usable variable, checks to see if the inventory.current_selected_item's name is the same as the interactObject.item_needed variable (string). sets the Player's "using_item" variable to true if it is. if the "item" button is pressed while "using_item" is true, a dialog box should pop up asking the player to confirm. if a quest item is used while "using_item" is false, a default error message should display - something like, "can't use that here."