can now use metal_pot on the sink
you get different dialog depending on if the sink is running or not.
next,  either replace the Metal_Pot with a brand new item, Metal_Pot_Full: (delete Metal_Pot from inventory, add Metal_Pot_Full to inventory)
or...
have the Metal_Pot have its own variable indiciating if its full or not.
more work to create a brand new Item for it, but there's less room for headaches that way, i think.
also should think about preventing the player from leaving home if that metal_pot is still in their inventory. since it is basically a useless object, it would clutter inventory space. either do it resident evil style where used items can be discarded, or make it so that the player MUST use the item properly to progress, which will automatically delete the item.