interaction between metal_pot and stove working
the big issue is having closing the dialog box consume the input event. as it is, closing a dialog box with the attack button launches an attack in addition to closing the dialog box, in one button press. consequently, closing the dialog box with the item button also uses an item. i'm not sure how to fix this.
also fixed the chargeUI not being shaded properly in the home.