making progress with dialog box
dialog box can now parse variables included in objects' dialog scripts.
buttons, to be used as responses in dialog, are created if the dialog script's dialog index (its array index) includes the keywords 'question' and 'options'.
'question' functions like the 'text' keyword: it writes text in the dialogbox. additionally, it looks for the 'options' keyword and creates buttons based on the corresponding options' dictionaries.
an example:

dialogBox.dialog_script = [
				{
					'text': "Hello.",
					'name': speaker
				},
				{
					'name': speaker,
					'question': 'Yes or no?',
					'options': [
						{ 'label': 'Yes', 'value': 'true'},
						{ 'label': 'No', 'value': 'false'}
					],
					'variable': 'answer'
				},
				{
					'name': speaker,
					'text': 'You said [answer].'
				}
			]