dialog 'checkpoint' system in place
there will be times when dialog boxes will ask to confirm a choice you've just made. if you want to change your mind, you can say no, and a 'checkpoint' will determine how far back in the dialog script to go. if you wish to re-answer a question, a value associated with the 'checkpoint' key will be subtracted from the dialog script's current index. usually, this value should be -3. pseudo example:
dialog index 1: "hello"
dialog index 2: "do you like A, B, or C?" ... display options
dialog index 3: "you picked [answer]."
dialog index 4:" is [answer] correct?" ... display options

if the answer is confirmed, the 'checkpoint' value is set to zero, meaning the dialog_index is unaffected and increments normally.
if the answer isn't correct, subtracts the 'checkpoint' value in the dialog_script from the dialog_index, which is usually -3.
load_dialog() increments the dialog index to 5, and 'checkpoint': '-3' reverts it to dialog_index 2, asking the question again.
if there are more lines of dialog between the question and the confirmation, the 'checkpoint' value can be adjusted accordingly.