testing dictionaries
biggest issue is now grasping how exactly the DialogBox parses what it finds in the dialog_script dictionaries, and then how to specifically access and parse that data into something usable elsewhere in the program (outside of the dialog box itself)
eg. talking with Skeleton instantiates a dialog box with the skeleton's dialog_script. i need to figure out a method to access other (selected) sections in the skeleton's dialog_script from the dialogBox itself AS WELL AS run functions in the skeleton's script from the dialogBox instantiated by the skeleton.
in other words, when an object creates a dialogBox, the dialogBox should get a reference to whatever it was that created it, but i don't know the best way of doing that.
thinking out loud:
perhaps if when a dialogBox is created, the object that instantiates it passes through its NodePath, which is stored as a variable in the dialogBox. this variable can then be used to run certain functions in the script of the object that created it (eg. skeleton). there are probably a million reasons why this won't work.