applying dirty indexing method
basically adding the 'skip' key to options, its value (integer) gets parsed as the number of indexes to skip in the dialog_script.
i have a feeling this would be better done with discrete dialog_scripts as opposed to jumping around to the right location within one. this will easily cause headaches down the line.
i think i will continue with this method until it works in an attempt to fully solidify my understanding of my own dialog system, THEN upgrade it to something cleaner.