tuning DialogBox behaviour
after almost 2 weeks of being stuck on various little problems, i have finally got the DialogBox to display and function as intended.
interactable objects have a dialog_script variable, which is an array of dictionaries. each dictionary refers to a single "page" or "box" of dialog, and may contain several key/value pairs that act as "events" to be interpreted by the dialogBox's event_handler.
currently used keys/events and their explanations:
'text': self-explanatory. writes the value as text in the dialog box.
'name': optional key for when speaking with another character. displays their name outside of the text box.
'question': the same as 'text', but handled as its own event which assumes there is an 'options' key that follows it.
'options': the value of 'options' is another array of dictionaries inside the dialog_script array of dictionaries. a Button is created for each dictionary. each embeded dictionary (ie., each 'option') may contain specific key/value pairs:
'label': sets the text of the button.
'skip': the number of dictionaries to skip over when loading the next dialog event, if that button ('option') is selected.
'value': not currently a very descriptive name - but when 'value' is '0', looks for a 'checkpoint' key and applies its integer value to the dialog_index.
'checkpoint': a negative integer. the number of indexes to "go back" in the dialog_script. useful for confirmation of options.
'variable': when an event has the 'variable' key, it lets the event_handler know to connect the button to the "_on_option_selected" function, and passes the value of 'variable' through to it. (the only 'variable' value currently is 'answer', which takes the value of the button's 'label' and applies it to a Global custom_variable that can be parsed to display the chosen options' text in the next dialog index whenever [answer] is typed in the 'text' key of a dialog_script's dictionary.)
'action': an event to perform an action, such as running a function on the object being interacted with, or ending the current dialog.
currently, the only two values associated with the 'action' key are:
'take_item', which advances dialog, parses/displays text, and runs the "acquire_item()" function on the object, and;
'end_dialog': runs the DialogBox's end_dialog() function, which unpauses the SceneTree and deletes the DialogBox instance.
more rigourous testing will need to be done to ensure the dialogBox doesn't break unexpededly and continues to function as intended.