beginning work on damage calculation / random variance
i do not wish to make a complex damage algorithm, but a number of things must be considered when designing one.
here is a very simple one, which i intend to try and modify as necessary:

damage = ((strength * modulator) - defense) * random

strength is the player's strength stat. the modulator is what the strength stat gets multiplied by, depending on the attack or maybe a buff. then, the enemy's defense stat is subtracted from the value. finally, random variance modulates the ultimate value; this value gets truncated into an integer, which is subtracted from the HP of what was attacked.
another perhaps "better" algorithm might be:

damage = attack * (100 / (100 + defense))

this has the advantage of "compressing" the final damage value, in addition to preventing attacks from doing 0 damage if the enemy's defense value exeeds the damage value.
basically, a lot of experimentation and planning will be required to make sure the game remains balances while the numbers feel good throughout.