created snow particle effect / modifying chargeUI
because i'm using parallax mirroring to transform (reposition) the particles' drawing area, it there is some unintended popin/popout of snowflakes when the camera scrolls vertically. horizontal scrolling is seamless, but i think due to the nature of the particles' animation (ie, it's vertical) it's not possible to "line up" in a cohesive way. the same thing happens with the rain, but i spent a lot of time modifying coordinates to mask the effect and the rain moves much faster, so it's harder to see. i really don't like how it feels so i will research writing my own shader, or perhaps won't use particle effects with parallax scrolling, because it doesn't seem to work very well.
also, made the chargeUI wheel flash red when charging if stamina < 25%. the closer to 0%, the faster the wheel flashes.