incorporated new stats / hit rate formula
all stats added to level pool (VIT, END, DEF, STR, DEX, SPD)
current stats can be viewed by pausing (spacebar)
enemy base accuracy default to 66
this value can change depending on what state the Player is in.
for example, during the Lv.1 charge attack, enemy_base_accuracy is reduced to 25. additionally, if you combo a Lv1. charge attack from a backdash, enemy_base_accuracy is reduced to 0!
currently, this is just a variable on the Player scene that gets modified and used as the first argument for enemy_hit_calculation. it feels strange that the enemy's base accuracy stat is based on what is essentially now a Player stat, but we'll see how/if this works in time.