modifying attacks / tuning gameplay / updating chargeUI
fixed a bug where queuing a charged attack during a backstep wouldn't activate
the hitbox of the flash attack has been tripled in length
the circular stamina bar now appears whenever stamina is set to anything below 100%, independently from the charge meters.
when stamina hits 0 and sweating is applied, the chargeUI disappears. an unintended cool effect is that while the stamina value is negative, the UI will "flash". this works as a visual indicator that you are depleted and unable to attack or dodge. i might even punish/debuff the player more if they ever reach this state.