player_hit_calculation integrated / other updates
updated bat & crow behaviour.
enemies that've been hit now check the player's accuracy and then run player_hit_calculation.
some example values: Player has 75 base_accuracy and 4 DEX. Bat has 8 Evasion. Crow has 12 evasion when it's on the ground, and 60 evasion while it's airborne.
with default DEX (4)...
you hit the bat 74% of the time.
you hit the Crow (ground) 71% of the time.
you hit the Crow (airborne) 23% of the time.
this is before any other attack/accuracy modifiers that may go on certain attacks.
with this, i've tried to make fighting the Crow more interesting: if you manage to hit it while it's airborne, it will land and be unable to attack for a moment. this includes if it is in its attack animation charging at you. if you manage to get it while its off guard, it's possible to stunlock combo it while it's on the ground.
also added a SoundPlayer singleton, for playing sound effects globally. currently used to play a sound when the player misses.
also various other gameplay tweaks. for example, the backdash animation decreases base_enemy_accuracy for the duration of the animation in addition to having iFrames. iFrames start at 1 and will scale up with speed, i think. maybe +0.1 frames per speed point.